Rename SequenceManagerConfig -> RoutingConfig
